// src/app/core/models/ms-audit-management/audit-program-report/audit-program-report.ts
export interface AuditReport {
    id: number;
    ms_audit_id: number;
    created_at: Date;
    updated_at: Date;
    created_by: CreatedBy;
    updated_by: any[];
    ms_audit_program_report_content: MSAuditAuditReportContent[];
}

export interface CreatedBy {
    id: number;
    first_name: string;
    last_name: string;
    email: string;
    mobile: null;
    image: Image;
    designation: string;
    department: string;
    status: Status;
}

export interface Image {
    title: string;
    url: string;
    token: string;
    size: number;
    ext: string;
    thumbnail_url: null;
}

export interface Status {
    id: number;
    title: Title[];
    label: string;
}

export interface Title {
    id: number;
    type: string;
    code: string;
    title: string;
    is_primary: number;
    is_rtl: number;
    created_at: Date;
    updated_at: Date;
    created_by: null;
    updated_by: null;
    status_id: number;
    pivot: TitlePivot;
}

export interface TitlePivot {
    status_id: number;
    language_id: number;
    title: string;
}

export interface MSAuditAuditReportContent {
    id: number;
    type: string;
    title: string;
    description: string;
    order: number;
    ms_audit_audit_report_id: number;
    created_by: number;
    created_at: Date;
    updated_by: null;
    updated_at: Date;
    section: string;
    ms_audit_audit_report_content_id: null;
    ms_audit_program_report_content_childrens: MSAuditAuditReportContentChildren[];
    list_of_non_conformities?: ListOfNonConformities;
}

export interface ListOfNonConformities {
    id: number;
    title: string;
    ms_audit_plan_id: number;
    start_date: Date;
    end_date: Date;
    created_at: Date;
    updated_at: Date;
    created_by: number;
    updated_by: null;
    ms_audit_findings: any[];
}

export interface MSAuditAuditReportContentChildren {
    id: number;
    type: string;
    title: string;
    description: string;
    order: number;
    ms_audit_audit_report_id: number;
    created_by: number;
    created_at: Date;
    updated_by: null;
    updated_at: Date;
    section: string;
    ms_audit_audit_report_content_id: number;
    audit_details?: AuditDetail[];
    ms_audit_audit_report_content_childrens: any[];
    audit_objectives?: AuditObjective[];
    audit_scopes?: Scope[];
    certificate_reference?: CertificateReference[];
    certificate_scope?: Scope[];
    audit_team?: AuditTeam;
    audit_agenda?: AuditAgenda;
}
// ...
export interface ActivityLogPaginationResponse {
    current_page: number;
    total: number;
    per_page: number;
    last_page: number;
    from: number;
    data: ActivityLogs[];
}

export interface ActivityLogs {
    id: number;
    activity: string;
    activity_type: string;
    created_at: string;
    created_by_department: string;
    created_by_designation: string;
    created_by_first_name: string;
    created_by_image_token: string;
    created_by_last_name: string;
    created_by_status: string;
    // data:Data;
    // detail:Detail[];
}